# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ByteProxy",
    "version": "__VERSION__",
    "description": "Extensible proxy for Discord, GitHub and other APIs with per-credential rate limiting.\n\n**Proxy authentication.** When `REQUIRE_AUTH_FOR_PROXY=true`, `/proxy/*` requires `PROXY_API_KEY` via `X-Api-Key`, `Authorization: Bearer`, or `?api_key=`. When `REQUIRE_AUTH_FOR_MANAGEMENT=true`, `/manage/*` requires `MANAGEMENT_API_KEY` the same way.\n\n**Upstream credentials.** ByteProxy holds no shared bot token. Callers authenticate to the upstream API themselves: send `X-Upstream-Authorization: Bot <token>` (or `Bearer <token>`), or send your normal `Authorization` header when the proxy key travels in `X-Api-Key`. Rate limits are tracked per upstream credential. Dynamic services may instead use a server-side `BYTEPROXY_TOKEN_*` environment variable.",
    "license": { "name": "AGPL-3.0-only", "url": "https://www.gnu.org/licenses/agpl-3.0.html" }
  },
...
  "components": {
    "securitySchemes": {
      "apiKeyHeader": { "type": "apiKey", "in": "header", "name": "X-Api-Key" },
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "API key" },
      "apiKeyQuery": { "type": "apiKey", "in": "query", "name": "api_key" }
    },
...
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" }, "service": { "type": "string" }, "path": { "type": "string" } },
        "required": ["error"]
      },
      "RateLimit": {
        "type": "object",
        "properties": { "maxRequests": { "type": "integer", "minimum": 1 }, "windowMs": { "type": "integer", "minimum": 1 } },
        "required": ["maxRequests", "windowMs"]
      },
      "Auth": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["bearer", "basic", "api-key", "bot"] },
          "tokenEnvVar": { "type": "string", "pattern": "^BYTEPROXY_TOKEN_[A-Z0-9_]+$", "description": "Optional server-side credential. Only BYTEPROXY_TOKEN_* variables are allowed." },
          "headerName": { "type": "string", "description": "Required for api-key auth." }
        },
        "required": ["type"]
      },
      "ServiceConfig": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 100 },
          "baseUrl": { "type": "string", "format": "uri", "description": "Absolute http(s) URL." },
          "headers": { "type": "object", "additionalProperties": { "type": "string" }, "description": "Static headers. Authorization, Cookie, Host and X-Api-Key can't be set here." },
          "rateLimit": { "$ref": "#/components/schemas/RateLimit" },
          "auth": { "$ref": "#/components/schemas/Auth" }
        },
        "required": ["name", "baseUrl"]
      },
      "ServiceSummary": {
        "type": "object",
        "properties": {
          "key": { "type": "string" },
          "name": { "type": "string" },
          "baseUrl": { "type": "string" },
          "builtin": { "type": "boolean" },
          "hasAuth": { "type": "boolean" },
          "authTokenConfigured": { "type": "boolean" },
          "rateLimit": { "oneOf": [{ "$ref": "#/components/schemas/RateLimit" }, { "type": "null" }] }
        }
      },
      "ServiceList": {
        "type": "object",
        "properties": { "services": { "type": "array", "items": { "$ref": "#/components/schemas/ServiceSummary" } }, "count": { "type": "integer" } }
      },
...
    },
    "responses": {
      "Unauthorized": { "description": "Missing or invalid API key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Json": { "description": "OK", "content": { "application/json": { "schema": { "type": "object" } } } }
    }
  },
  "paths": {
...
    "/manage/services": {
      "get": {
        "tags": ["Management"],
        "summary": "List all services",
        "operationId": "manageListServices",
        "security": [{ "apiKeyHeader": [] }, { "bearerAuth": [] }, { "apiKeyQuery": [] }],
        "responses": { "200": { "description": "Services", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ServiceList" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      },
      "post": {
        "tags": ["Management"],
        "summary": "Register a service",
        "description": "Adds a service and saves it to SERVICES_FILE so it survives restarts.",
        "operationId": "manageAddService",
        "security": [{ "apiKeyHeader": [] }, { "bearerAuth": [] }, { "apiKeyQuery": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "key": { "type": "string", "pattern": "^[a-z0-9][a-z0-9_-]{0,31}$" }, "config": { "$ref": "#/components/schemas/ServiceConfig" } }, "required": ["key", "config"] },
              "example": { "key": "status", "config": { "name": "Status API", "baseUrl": "https://status.example.com/api/", "rateLimit": { "maxRequests": 60, "windowMs": 60000 } } }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Json" },
          "400": { "description": "Invalid configuration", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "description": "Key already exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}